backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
asyncpg
psycopg2-binary
python-jose
passlib
bcrypt<4.0
authlib
itsdangerous
pydantic[email]
python-dotenv
httpx
langchain
langgraph
langchain-groq
tenacity